Rewrite and rename `gen:openapi` script
