+++ frontend/package.json
@@ -23,7 +23,7 @@
     "build:ladle": "ladle build",
     "start:msw": "cross-env API_MODE=mock vite",
     "gen:icons": "node scripts/gen_icons.js",
-    "gen:openapi": "vite-node ./scripts/gen_openapi_types.ts",
+    "gen:openapi": "vite-node scripts/gen_openapi.ts",
     "gen:po": "vite-node scripts/gen_po.js",
     "gen:translation-json": "vite-node scripts/gen_translation-json.js",
     "serve": "npm run build:msw && npx serve dist"
